anyida-scraper/requirements.txt
celery
redis
seleniumbase
playwright
beautifulsoup4
requests
python-dotenv
tenacity
fake-useragent
curl_cffi
